Code cleanup + Unit tests ---  ### Summary of work done in the commit:  - Added new unit tests.  - Code cleanup.  - Plan update. ---  ### TODOs before issue close:  - Finish implementing tests and full prototype pipeline.
See #6
