+++ timelog.md
@@ -160,3 +160,9 @@
 ## 8 Oct 2020
 
 - _1 hour_ - Working on lipid kinetics tests. Added label property to lipids, need to add this ability on Lipid info page.
+
+## 10 Oct 2020
+
+- _1.5 hours_ - Writing tests for isotope pipeline - Have successfully written a test for get_max_mass which attempts to get the max value of a specific scan read from TOPPView.
+
+## 11 Oct 2020
